+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 - Add `SegmentTimeline` element to `SegmentList` elements (from @erik-moqvist). 
 
+- Add definition for `BitstreamSwitching` elements to `SegmentTemplate` and `SegmentList` nodes.
+
 - Fix type of `@bitstreamSwitching` attribute on `SegmentTemplate` elements (xs:string rather than
   xs:bool as for all other uses of the `@bitstreamSwitching` attribute).